Add last few sub view models
